improve debiso build

+++ config/archiso/bootup.sh
@@ -3,9 +3,11 @@
 
 # stop and disable reflector to be able to change the pacman mirrorlist
 # run pacman init when everything is ready
-systemctl stop reflector.service reflector.timer pacman-init.service
-systemctl disable reflector.service reflector.timer pacman-init.service
-systemctl mask reflector.service reflector.timer
+if [ -f /usr/lib/systemd/system/reflector.service ]; then
+  systemctl stop reflector.service reflector.timer pacman-init.service
+  systemctl disable reflector.service reflector.timer pacman-init.service
+  systemctl mask reflector.service reflector.timer
+fi
 
 # improve boot time by not waiting for ntp
 systemctl stop systemd-time-wait-sync.service
